Generic Projects Endpoint
Added a new possibility to request projects by slug, namespace or namespace/slug

1) To request projects using slug use /api/v1/data-projects/slug/{projectslug}
2) To request projects using namespace use /api/v1/data-projects/namespace/{namespace}
3) To request a project using namespace/slug use /api/v1/data-projects/{namespace}/{slug}

frontend#424

+++ mlreef-rest/src/main/kotlin/com/mlreef/rest/external_api/gitlab/requests.kt
@@ -90,7 +90,7 @@ internal class GitlabCreateProjectRequest(
     val ciConfigPath: String = "mlreef.yml",
     val description: String = "",
     val defaultBranch: String = "",
-    val namespace_id: Int? = null
+    val namespace_id: Long? = null
 ) : Serializable
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
